rolemodel part 변경

+++ app/src/main/java/com/fournineseven/dietstock/model/getRanking/RankingResult.java
@@ -16,9 +16,9 @@ public class RankingResult {
 
     @SerializedName("kcal")
     @Expose
-    private int kcal;
+    private Float kcal;
 
-    public RankingResult(int user_no, String name, int kcal){
+    public RankingResult(int user_no, String name, Float kcal){
         this.user_no = user_no;
         this.name = name;
         this.kcal = kcal;
@@ -32,11 +32,11 @@ public class RankingResult {
         this.name = name;
     }
 
-    public int getKcal() {
+    public Float getKcal() {
         return kcal;
     }
 
-    public void setKcal(int kcal) {
+    public void setKcal(Float kcal) {
         this.kcal = kcal;
     }
